added area of polygon feature
Running extract_features will display some functionality including a polygon between the leading 3 vertices. An area is calculated too and we can start to pull more metrics out of the leading three droplets to build a matrix of features and perform kNN. Trying to write out a video of the animation but having trouble opening the file on my machine.